small refactor
change package and classes names
