+++ handlers/zuora-reports/src/test/scala/com/gu/zuora/reports/QuerierEffectsTest.scala
@@ -3,9 +3,10 @@ package com.gu.zuora.reports.aqua
 import com.gu.effects.{RawEffects, S3ConfigLoad}
 import com.gu.test.EffectsTest
 import com.gu.util.config.{LoadConfig, Stage}
-import com.gu.zuora.reports.AquaLambda.StepsConfig
+import com.gu.zuora.reports.ReportsLambda.StepsConfig
+import com.gu.zuora.reports.{Querier, QuerierRequest, QuerierResponse, Query}
 import org.scalatest.{FlatSpec, Matchers}
-import scalaz.{-\/, \/}
+import scalaz.-\/
 import scalaz.syntax.std.either._
 
 //TODO SEE HOW TO DO EFFECT TESTS FOR REPORT QUERIES SINCE THEY ALL DEPEND ON THIS ONE
