+++ handlers/zuora-reports/src/test/scala/com/gu/zuora/reports/aqua/AquaRequestTest.scala
@@ -1,5 +1,6 @@
 package com.gu.zuora.reports.aqua
 
+import com.gu.zuora.reports.{AquaQuery, AquaQueryRequest}
 import org.scalatest._
 import play.api.libs.json.Json
 import org.scalatest.Matchers._
